% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.r
\name{boy.surface}
\alias{boy.surface}
\title{Boy Surface}
\usage{
boy.surface(n = 10000)
}
\arguments{
\item{n}{number of points}
}
\description{
A function to produce a Boy Surface.
}
\examples{
## Generates a Boy Surface
boy.surface(n = 1000)

}
\author{
Barret Schloerke
}
\references{
\url{http://barretschloerke.com/GeoZoo/mobius/other/}
}
\keyword{dynamic}

