% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{cross.cap}
\alias{cross.cap}
\title{Cross Cap}
\usage{
cross.cap(n = 10000)
}
\arguments{
\item{n}{number of points}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object (null)}
}
\description{
A function to generate a cross cap
}
\examples{
## Generates a Cross Cap
cross.cap( n = 1000 )

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/mobius/other/}
}
\keyword{dynamic}

