% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube.dotline}
\alias{cube.dotline}
\title{Cube with points along the wire frame}
\usage{
cube.dotline(p = 3)
}
\arguments{
\item{p}{dimension of object}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object}
}
\description{
A function to generate a cube with points on its face
}
\examples{
## Generates a cube with points along its wire frame
cube.dotline(p = 3)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/cube/}
}
\keyword{dynamic}

