% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube.face.grid}
\alias{cube.face.grid}
\title{Cube with equidistant points on the 'face'}
\usage{
cube.face.grid(p = 3, n = 8)
}
\arguments{
\item{p}{dimension of object}

\item{n}{length of number of points in each dimension}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object}
}
\description{
A function to generate a cube with equidistant points on its face
}
\examples{
## Generates a cube with points on its face
cube.face(p = 3)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/cube/}
}
\keyword{dynamic}

