% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube.face.random}
\alias{cube.face}
\alias{cube.face.random}
\title{Cube with random points on the 'face'}
\usage{
cube.face.random(p = 3, n = 850 * 2^(p - 1))

cube.face(p = 3, n = 850 * 2^(p - 1))
}
\arguments{
\item{p}{dimension of object}

\item{n}{number of points on each face}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object}
}
\description{
A function to generate a cube with random points on its face
}
\examples{
## Generates a cube with points on its face
cube.face(p = 3)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/cube/}
}
\keyword{dynamic}

