\name{cube.iterate}
\alias{cube.iterate}
\title{Cube}
\usage{
  cube.iterate(p = 3)
}
\arguments{
  \item{p}{dimension of object}
}
\value{
  \item{points }{location of points} \item{edges }{edges of
  the object}
}
\description{
  A function generate a cube with vertices and a wire frame
}
\examples{
## Generates a cube
cube.iterate(p = 3)
}
\author{
  Barret Schloerke
}
\references{
  \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/cube/}
}
\keyword{dynamic}

