% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{f_composition}
\alias{f_composition}
\title{f_composition}
\usage{
f_composition(data)
}
\arguments{
\item{data}{object}
}
\value{
\item{data }{points in (d-1)-dimensional space}
}
\description{
Function to take a d-dimensional compositional data set and 
transform it using a Helmert transformation into (p-1)-space, 
where it lives. Mostly internally used, but could be useful 
for setting up new compositional data.
}
\author{
Di Cook
}
\references{
\url{http://schloerke.github.io/geozoo/simplices/}
}

