% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{f_helmert}
\alias{f_helmert}
\title{f_helmert}
\usage{
f_helmert(d)
}
\arguments{
\item{d}{object}
}
\value{
\item{helmert }{transformation matrix}
}
\description{
Function to set up a Helmert transformation of a (d-1)-dimensional 
shape in p-space down into its (p-1)-space. Mostly internally used, 
but could be useful for setting up new compositional data.
}
\author{
Di Cook
}
\references{
\url{http://schloerke.github.io/geozoo/simplices/}
}

