% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius.R
\name{mobius.experiment}
\alias{mobius.experiment}
\title{Mobius Experiment}
\usage{
mobius.experiment(p = 5, n = 10000)
}
\arguments{
\item{p}{dimension of object.  (5)}

\item{n}{number of points}
}
\description{
A function to generate a 5-D mobius strip in the third dimension.
}
\examples{
## Generates a mobius strip.
mobius.experiment(5, n = 1000)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/mobius/mobius/}
}
\keyword{dynamic}

