% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.geozoo}
\alias{print.geozoo}
\title{Print}
\usage{
\method{print}{geozoo}(x, ...)
}
\arguments{
\item{x}{geozoo object}

\item{...}{other arguements}
}
\description{
Prints geozoo objects with tourr or prints them
}
\examples{
\dontrun{
example(boy.surface)
example(cube.face)
}
}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/}
}
\keyword{dynamic}

