% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.geozooNoScale}
\alias{print.geozooNoScale}
\title{Print Without Rescale}
\usage{
\method{print}{geozooNoScale}(x, ...)
}
\arguments{
\item{x}{geozoo object}

\item{...}{other arguements}
}
\description{
Prints objects without rescaling them to 0, 1 in each dim
}
\examples{
\dontrun{
torus()
}
}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/}
}
\keyword{dynamic}

