\name{boy.surface}
\alias{boy.surface}
\title{Boy Surface}
\usage{
  boy.surface(n = 10000)
}
\arguments{
  \item{n}{number of points}
}
\description{
  A function to produce a Boy Surface.
}
\examples{
## Generates a Boy Surface
boy.surface(n = 1000)
}
\author{
  Barret Schloerke
}
\references{
  \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/mobius/other/}
}
\keyword{dynamic}

