\name{cross.cap}
\alias{cross.cap}
\title{Cross Cap}
\usage{
  cross.cap(n = 10000)
}
\arguments{
  \item{n}{number of points}
}
\value{
  \item{points }{location of points} \item{edges }{edges of
  the object (null)}
}
\description{
  A function to generate a cross cap
}
\examples{
## Generates a Cross Cap
cross.cap( n = 1000 )
}
\author{
  Barret Schloerke
}
\references{
  \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/mobius/other/}
}
\keyword{dynamic}

