% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.polytope.R
\name{cross.polytope}
\alias{cross.polytope}
\title{Cross Polytope}
\usage{
cross.polytope(p = 3)
}
\arguments{
\item{p}{dimension of object}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object}
}
\description{
A function generate a cross polytope, cube dual, with vertices and a wire frame.
}
\examples{
# Generates a 3-D Cross Polytope
cross.polytope(p = 3)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/cube/}
}
\keyword{dynamic}

