% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube.solid.grid}
\alias{cube.solid.grid}
\title{Equidistant Solid Cube}
\usage{
cube.solid.grid(p = 3, n = 8)
}
\arguments{
\item{p}{dimension of object}

\item{n}{length of number of points in each dimension}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object}
}
\description{
A function to generate a solid cube with equidistant points
}
\examples{
## Generates a solid cube with equidistant points
cube.solid.grid(p = 3, n = 8)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/cube/}
}
\keyword{dynamic}

