% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobius.R
\name{mobius}
\alias{mobius}
\title{Mobius}
\usage{
mobius(p = 3, n = 10000)
}
\arguments{
\item{p}{dimension of object.  (3)}

\item{n}{number of points}
}
\description{
A function to generate a mobius strip in the third or fourth dimension.
}
\examples{
## Generates a mobius strip.
mobius(3, n = 1000)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/mobius/mobius/}
}
\keyword{dynamic}

