\name{print.geozoo}
\alias{print.geozoo}
\title{Print}
\usage{
  \method{print}{geozoo} (x, ...)
}
\arguments{
  \item{x}{geozoo object}

  \item{...}{other arguements}
}
\description{
  Prints geozoo objects with tourr or prints them
}
\examples{
\dontrun{
example(boy.surface)
example(cube.face)
}
}
\author{
  Barret Schloerke
}
\references{
  \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/}
}
\keyword{dynamic}

