\name{print.geozoo}
\alias{print.geozoo}
\title{ Print }
\description{
Prints GeoZoo objects with tourr or prints them
}
\arguments{
  \item{x}{ geozoo object}
  \item{...}{ other arguements}
}
\references{ \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/} }
\author{ Barret Schloerke }
\examples{
example(boy.surface)
example(cube.face)
}
\keyword{ dynamic }