\name{simplex}
\alias{simplex}
\title{Simplex}
\usage{
  simplex(p = 3)
}
\arguments{
  \item{p}{dimension of object}
}
\value{
  \item{points }{location of points} \item{edges }{edges of
  the object (null)}
}
\description{
  A function to generate a simplex
}
\examples{
## Generates a simplex
simplex(p = 3)
}
\author{
  Barret Schloerke
}
\references{
  \url{http://streaming.stat.iastate.edu/~dicook/geometric-data/simplex/}
}
\keyword{dynamic}

