% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.R
\name{torus}
\alias{torus}
\title{Torus}
\usage{
torus(p = 3, n = 10000, radius = 2^((p - 2):0))
}
\arguments{
\item{p}{dimension of object}

\item{n}{number of points}

\item{radius}{radiuses of the torus, set from largest to smallest}
}
\value{
\item{points }{location of points}
 \item{edges }{edges of the object (null)}
}
\description{
A function to generate a torus in any dimension
}
\examples{
## Generates a torus
torus(p = 3, n = 1000)

}
\author{
Barret Schloerke
}
\references{
\url{http://schloerke.github.io/geozoo/mobius/torus/}
}
\keyword{dynamic}

