\name{write.xml}
\alias{write.xml}
\title{Write XML File}
\usage{
  write.xml(object.function, filepath, title)
}
\arguments{
  \item{object.function}{i.e. cube() or sphere()}

  \item{filepath}{where you would like to save the file in
  quotes}

  \item{title}{title of the file in quotes}
}
\description{
  A function that allows the user to write an object into
  an .xml.
}
\author{
  Barret Schloerke
}
\keyword{dynamic}

